+++ NEWS.md
@@ -1,3 +1,7 @@
+# 2025-11-21
+
+- Generic boosting
+
 # 2025-11-04
 
 - Initial version